Use mitm-receiver
This changes things to use mitm-receiver. When using mitm-receiver,
protos will not be POSTed to MAD's MITMReceiver.

Instead, mitm-receiver will queue up protos and workers will poll mitm-receiver
for them.

Also, we query mitm-receiver for the most recent proto received along
with location information, etc, to know when to move to next location.